requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
google-api-python-client==2.86.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0